+++ src/rules/index.ts
@@ -80,7 +80,6 @@ import { leaZeroAddress } from "./optimization/lea-zero-address.js";
 import { moveImmediateWordComplement, moveImmediateSwap } from "./optimization/move-immediate-more-synthesis.js";
 import { longShiftSequence } from "./optimization/long-shift-sequences.js";
 import { moveImmediateAddressToLea, moveAddressThenAddToLea } from "./optimization/movea-lea.js";
-import { preferLeaForAddressSymbol } from "./optimization/lea-address-symbol.js";
 import { atariTrapStackCleanup } from "./platform/atari/trap-stack-cleanup.js";
 import { amigaBitMaskConstants } from "./platform/amiga/bit-mask-constants.js";
 import { cancelMultiplePredecrementMoves } from "./optimization/cancel-multiple-predecrement.js";
@@ -223,7 +222,6 @@ export {
   longShiftSequence,
   moveImmediateAddressToLea,
   moveAddressThenAddToLea,
-  preferLeaForAddressSymbol,
   cancelMultiplePredecrementMoves,
   cancelStackPeaSequence,
   multiplyLongByOne,
@@ -314,7 +312,6 @@ export const defaultRules: readonly Rule[] = [
   longShiftSequence,
   moveImmediateAddressToLea,
   moveAddressThenAddToLea,
-  preferLeaForAddressSymbol,
   cancelMultiplePredecrementMoves,
   moveImmediateWordComplement,
   moveImmediateSwap,